Player Movement + Hover
In SampleScene there's now a Player that uses Unity's new input system to move around, and jump. With options in the Inspector window. If one of the variables is required for your scripting needs, just make it public so your scripts can access those specific variables. Hovering also works as well, by holding down the Spacebar.