Added 3 Achievements
You get all three by:
Killing your first enemy
Dying to an enemy
Finding the secret passageway (it's not that secret so you should be able to spawn in, go right, past the skeleton and it's there)

There's one issue and its that the achievements don't save when you load into another scene. So replaying the first level gets you the same achievements. If this is a huge problem, then I'll fix it later this week.